refactor:TOP-93 TermsDao를 TermsService로 변경
 - Remote, Local 변경 가능성을 두고 Service로 변경
