+++ data/src/main/java/com/tht/tht/data/di/DaoModule.kt
@@ -1,11 +1,9 @@
 package com.tht.tht.data.di
 
-import com.tht.tht.data.local.dao.TokenDao
-import com.tht.tht.data.local.dao.TokenDaoImpl
 import com.tht.tht.data.local.dao.SignupUserDao
 import com.tht.tht.data.local.dao.SignupUserDaoImpl
-import com.tht.tht.data.local.dao.TermsDao
-import com.tht.tht.data.local.dao.TermsDaoImpl
+import com.tht.tht.data.local.dao.TokenDao
+import com.tht.tht.data.local.dao.TokenDaoImpl
 import com.tht.tht.data.local.dao.topic.DailyTopicDao
 import com.tht.tht.data.local.dao.topic.DailyTopicDaoImpl
 import dagger.Binds
@@ -17,10 +15,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class DaoModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindTermsDao(impl: TermsDaoImpl): TermsDao
 
     @Binds
     @Singleton
