--- data/src/main/java/com/tht/tht/data/local/dao/TermsDao.kt
@@ -1,7 +1,0 @@
-package com.tht.tht.data.local.dao
-
-import com.tht.tht.data.local.entity.TermsEntity
-
-interface TermsDao {
-    suspend fun fetchTerms(): TermsEntity
-}
