+++ data/src/main/java/com/tht/tht/data/local/service/TermsService.kt
@@ -1,0 +1,7 @@
+package com.tht.tht.data.local.service
+
+import com.tht.tht.data.local.entity.TermsEntity
+
+interface TermsService {
+    fun fetchTerms(): TermsEntity
+}
