+++ data/src/test/java/com/tht/tht/data/datasource/TermsDataSourceImplTest.kt
@@ -1,6 +1,6 @@
 package com.tht.tht.data.datasource
 
-import com.tht.tht.data.local.dao.TermsDao
+import com.tht.tht.data.local.service.TermsService
 import com.tht.tht.data.local.datasource.TermsDataSourceImpl
 import com.tht.tht.data.local.entity.TermsEntity
 import io.mockk.coEvery
@@ -19,7 +19,7 @@ import org.junit.Test
 internal class TermsDataSourceImplTest {
 
     private lateinit var dataSource: TermsDataSourceImpl
-    private lateinit var dao: TermsDao
+    private lateinit var dao: TermsService
     private val testDispatcher = StandardTestDispatcher(TestCoroutineScheduler())
 
     @Before
